refactor: move to src/
