+++ src/jump_profiling_recipe/correct/corrections.py
@@ -21,8 +21,8 @@ import numpy as np
 from statsmodels.formula.api import ols
 import logging
 from tqdm.auto import tqdm
-from preprocessing.utils import validate_columns as validate_columns
-from preprocessing.metadata import get_feature_columns, get_metadata_columns
+from ..preprocessing.utils import validate_columns as validate_columns
+from ..preprocessing.metadata import get_feature_columns, get_metadata_columns
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
