+++ src/jump_profiling_recipe/correct/format_check.py
@@ -6,7 +6,7 @@ import os
 import pandas as pd
 import numpy as np
 import logging
-from preprocessing import io
+from ..preprocessing import io
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
