+++ src/jump_profiling_recipe/preprocessing/feature_selection.py
@@ -4,7 +4,7 @@ Functions for feature selection
 
 import logging
 import pandas as pd
-from pycytominer.operations import correlation_threshold, variance_threshold
+from ..pycytominer.operations import correlation_threshold, variance_threshold
 from .metadata import get_feature_columns
 
 logger = logging.getLogger(__name__)
