+++ src/jump_profiling_recipe/preprocessing/sphering.py
@@ -13,7 +13,7 @@ import pandas as pd
 
 from .io import merge_parquet, split_parquet
 from .metadata import NEGCON_CODES
-from pycytominer.operations import Spherize
+from ..pycytominer.operations import Spherize
 
 
 def generate_log_uniform_samples(
