+++ src/tests/fixtures/inputs/cell_counts/crispr_cell_counts.csv.gz
