+++ src/tests/fixtures/inputs/config/crispr_trimmed.json
