+++ src/tests/fixtures/inputs/metadata/plate.csv.gz
