+++ src/tests/scripts/copy_parquet_fixture.py
