+++ src/tests/test_pipeline_integration.py
